refactor: rename and move react-mitt in Decorators
